requirements.txt
ollama
requests
beautifulsoup4
google-genai
python-dotenv
jinja2
msal
markdown